small modifications

+++ src/main/java/it/polimi/db/rest/DeletionController.java
@@ -2,17 +2,10 @@ package it.polimi.db.rest;
 
 import java.io.IOException;
 import java.sql.Date;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -24,15 +17,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import it.polimi.db.entity.Answer;
-import it.polimi.db.entity.Question;
 import it.polimi.db.entity.Questionnaire;
 import it.polimi.db.entity.Statistic;
-import it.polimi.db.entity.User;
-import it.polimi.db.form.AnswersForm;
 import it.polimi.db.service.AnswerService;
 import it.polimi.db.service.QuestionnaireService;
 import it.polimi.db.service.StatisticService;
-import it.polimi.db.service.UserService;
 
 @Controller
 @RequestMapping("/deletion")
@@ -54,9 +43,8 @@ public class DeletionController {
 	
 	@ModelAttribute("yesterdaysQuestionnaire")
 	public Questionnaire getYesterdaysQuestionnaire() {
-		Optional<Questionnaire> yesterdays = questionnaireService.findByDate(new Date(System.currentTimeMillis()));//-24*60*60*1000));
-		if (yesterdays.isPresent()) 
-			 return yesterdays.get();
+		Optional<Questionnaire> yesterdays = questionnaireService.findByDate(new Date(System.currentTimeMillis()-24*60*60*1000));
+		if (yesterdays.isPresent()) return yesterdays.get();
 		return null;
 	}
 	
@@ -65,26 +53,27 @@ public class DeletionController {
 			BindingResult bindingResult, Model model, HttpServletRequest req) throws IOException {
 
 		if (method.equals("Delete")) {
-			Optional<Questionnaire> yesterdays = questionnaireService.findByDate(new Date(System.currentTimeMillis()));//-24*60*60*1000));
+			Optional<Questionnaire> yesterdays = questionnaireService.findByDate(new Date(System.currentTimeMillis()-24*60*60*1000));
 			if (yesterdays.isPresent()) {
-				 Questionnaire questionnaire = yesterdays.get();
-				 questionnaireService.removeQuestionnaire(questionnaire);
-				 List<Answer> answers = answerService.findByQuestionnaire(questionnaire.getId());
-				 for(Answer a : answers) {
-					 answerService.removeAnswer(a);
-				 }
-				 List<Statistic> stats = statService.listAll();
-				 for(Statistic s : stats) {
-					 if(s.getQuestionnaireId()==questionnaire.getId())
-						 statService.removeStatistic(s);
-				 }
+				Questionnaire questionnaire = yesterdays.get();
+				questionnaireService.removeQuestionnaire(questionnaire);
+				List<Answer> answers = answerService.findByQuestionnaire(questionnaire.getId());
 				 
-				 model.addAttribute("msg", "Thank you for deleting the questionnaire!");
-			}
-			else {
+				for(Answer a : answers) {
+					answerService.removeAnswer(a);
+				}
+				 
+				List<Statistic> stats = statService.listAll();
+				for(Statistic s : stats) {
+					if(s.getQuestionnaireId()==questionnaire.getId()) statService.removeStatistic(s);
+				}
+				 
+				model.addAttribute("msg", "Thank you for deleting the questionnaire!");
+			} else {
 				model.addAttribute("msg", "Deletion of the questionnaire unsuccessful!");
 			}
 		}
+		
 		return "uploadstatus";
 	}
